# src/main/java/com/example/boxticketingwebapi/controller/EventController.java
package com.example.boxticketingwebapi.controller;

import com.example.boxticketingwebapi.model.EventModel;
import com.example.boxticketingwebapi.service.EventService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/v1/events")
public class EventController {


    @Autowired
    private EventService eventService;

    @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.OK)
    public @ResponseBody /*Inserts returned value in reponse body*/ List<EventModel> getEvents() throws Exception {
        return this.eventService.getAllEvents();
    }

    @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.OK)
    public @ResponseBody /*Inserts returned value in reponse body*/ EventModel getEvent(@PathVariable(value = "id") Integer eventId) throws Exception {
        EventModel event = this.eventService.getEventDetails(eventId);
        event.add(WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(this.getClass()).addEvent(new EventModel())).withRel("addEvent"));
        event.add(WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(this.getClass()).getEvents()).withRel("getEvents"));
        return event;
    }

    @PostMapping(value = "/addEvent" , consumes = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.CREATED)
    public EventModel addEvent(@RequestBody EventModel event) {
        return this.eventService.saveEvent(event);
    }

    @DeleteMapping(value = "/deleteEvent/{id}")
    @ResponseStatus(HttpStatus.OK)
    public String deleteEvent(@PathVariable(value = "id") Integer eventId) {
        return this.eventService.deleteEvent(eventId);
    }

}